+++ cmd/stage.go
@@ -2,13 +2,14 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/go-imsto/imsto/storage"
-	"io"
 	"log"
 	"net/http"
 	"runtime"
 	"strconv"
 	"time"
+
+	"github.com/go-imsto/imsto/config"
+	"github.com/go-imsto/imsto/web"
 )
 
 var cmdStage = &Command{
@@ -29,43 +30,6 @@ func init() {
 	cmdStage.Run = runStage
 }
 
-func StageHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Add("X-Server", "IMSTO STAGE")
-
-	item, err := storage.LoadPath(r.URL.Path)
-
-	if err != nil {
-		logger().Warnw("loadPath fail", "ref", r.Referer(), "err", err)
-		switch err.(type) {
-		case *storage.HttpError:
-			ie := err.(*storage.HttpError)
-			if ie.Code == 302 {
-				// log.Printf("redirect to %s", ie.Path)
-				http.Redirect(w, r, ie.Path, ie.Code)
-				return
-			}
-			// w.WriteHeader(ie.Code)
-			http.Error(w, ie.Text, ie.Code)
-			return
-		}
-		writeJsonError(w, r, err)
-
-		return
-	}
-
-	// log.Print(item)
-
-	c := func(file io.ReadSeeker) {
-		http.ServeContent(w, r, item.Name(), item.Modified(), file)
-	}
-	err = item.Walk(c)
-	if err != nil {
-		logger().Warnw("item walk fail", "item", item, "err", err)
-		writeJsonError(w, r, err)
-		return
-	}
-}
-
 func runStage(args []string) bool {
 	if *sMaxCpu < 1 {
 		*sMaxCpu = runtime.NumCPU()
@@ -73,10 +37,9 @@ func runStage(args []string) bool {
 	runtime.GOMAXPROCS(*sMaxCpu)
 
 	var e error
-	http.HandleFunc("/", StageHandler)
+	http.HandleFunc("/", web.StageHandler)
 
-	// log.Print("Start Stage service ", VERSION, " at port ", strconv.Itoa(*sport))
-	str := fmt.Sprintf("Start Stage service %s at port %d", VERSION, *sport)
+	str := fmt.Sprintf("Start Stage service %s at port %d", config.Version, *sport)
 	fmt.Println(str)
 	log.Print(str)
 	srv := &http.Server{
