+++ storage/logger.go
@@ -1,0 +1,9 @@
+package storage
+
+import (
+	zlog "github.com/go-imsto/imsto/log"
+)
+
+func logger() zlog.Logger {
+	return zlog.Get()
+}
